42api-scraping/requirements.txt
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
requests-cache==0.9.3
bcrypt==3.2.0
pymongo==4.1.1
python-dotenv==0.20.0
